mcp/requirements.txt
mcp[cli]>=1.12.0
httpx>=0.27.0
python-dotenv>=1.0.0
